+++ src/commands/automaticProofreader.ts
@@ -1,17 +1,6 @@
 import * as vscode from 'vscode';
 import * as autoProofreader from '../commons/automaticProofreader';
 
-// const sum = (line: string) =>{
-// 	let sum = 0;
-// 	for (let index = 0; index < line.length; index++) {
-// 		sum=sum+line.toUpperCase().charCodeAt(index);
-// 		if(sum > 255){
-// 			sum = sum % 256;
-// 		}
-// 	}
-// 	return sum;
-// };
-
 export function automaticProofreader () {
 	let checksum = -1;
 	const editor = vscode.window.activeTextEditor;
@@ -27,12 +16,3 @@ export function automaticProofreader () {
 	}
 	return checksum;
 };
-
-// export function automaticProofreaderSilent (line: string) {
-// 	let checksum = -1;
-// 	if (line) {
-// 		const trimmedLine = line.trim().replace(/ /g,"");
-// 		checksum = sum(trimmedLine);
-// 	}
-// 	return checksum;
-// };