+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "c64basicv2" extension will be documented in this file.
 
+## [0.6.2] - 01/05/2022
+
+- [Diagnostics] Added check line start with a number
+- [Automatic Proofreader] Disabled by default
+- [Automatic Proofreader] Added control character support
+
 ## [0.6.1] - 24/04/2022
 
 - [Diagnostics] Added "enableDiagnostics" configuration property